Refactor: Article interface 네임 변경
