--- interface/article.ts
@@ -1,6 +1,0 @@
-export interface ArticleInteface {
-  [x: string]: any;
-  title: string;
-  tag: string;
-  content: string;
-}
