+++ src/components/templates/ArticleList/ArticleList.tsx
@@ -1,7 +1,7 @@
 import { useState, useEffect } from "react";
 import { findArticles } from "src/api/article";
 import { ArticleBox } from "src/components/UI//molecules";
-import { ArticleInteface } from "interface";
+import { Article as ArticleProps } from "src/types";
 
 const ArticleList = () => {
   const [articles, setArticles] = useState([]);
@@ -18,7 +18,7 @@ const ArticleList = () => {
   return (
     <>
       <div>
-        {articles.map((article: ArticleInteface) => (
+        {articles.map((article: ArticleProps) => (
           <div key={article._id}>
             <ArticleBox article={article} />
           </div>
