+++ src/pages/articles/[id].tsx
@@ -8,7 +8,7 @@ import { serialize } from "next-mdx-remote/serialize";
 import { MDXRemote } from "next-mdx-remote";
 import ArticleModel from "src/pages/api/models/article";
 import connectMongo from "src/pages/api/utils/connectMongo";
-import { ArticleInteface } from "interface";
+import { Article as ArticleProps } from "src/types";
 
 const Article = ({
   title,
@@ -28,7 +28,7 @@ export const getStaticPaths: GetStaticPaths = async () => {
   await connectMongo();
   const res = await ArticleModel.find();
   const articles = JSON.parse(JSON.stringify(res));
-  const paths = articles.map((article: ArticleInteface) => ({
+  const paths = articles.map((article: ArticleProps) => ({
     params: { id: article?._id },
   }));
