+++ src/types/article.ts
@@ -1,0 +1,6 @@
+export interface Article {
+  [key: string]: string;
+  title: string;
+  tag: string;
+  content: string;
+}
